+++ src/main/java/gregtech/api/graphs/consumers/NodeEnergyConnected.java
@@ -16,7 +16,7 @@ public class NodeEnergyConnected extends ConsumerNode {
     }
 
     @Override
-    public int injectEnergy(int aVoltage, int aMaxAmps) {
+    public int injectEnergy(long aVoltage, long aMaxAmps) {
         return (int) ((IEnergyConnected)mTileEntity).injectEnergyUnits(mSide,aVoltage,aMaxAmps);
     }
 }
