+++ src/main/java/gregtech/api/graphs/consumers/NodeEnergySink.java
@@ -6,7 +6,7 @@ import net.minecraftforge.common.util.ForgeDirection;
 
 import java.util.ArrayList;
 
-//consumer for IC2 machines
+// consumer for IC2 machines
 public class NodeEnergySink extends ConsumerNode {
     public NodeEnergySink(int nodeValue, IEnergySink tileEntity, byte side, ArrayList<ConsumerNode> consumers) {
         super(nodeValue, (TileEntity) tileEntity, side, consumers);
@@ -18,7 +18,7 @@ public class NodeEnergySink extends ConsumerNode {
     }
 
     @Override
-    public int injectEnergy(int aVoltage, int aMaxAmps) {
+    public int injectEnergy(long aVoltage, long aMaxAmps) {
         int tUsedAmps = 0;
         while (aMaxAmps > tUsedAmps && ((IEnergySink) mTileEntity).getDemandedEnergy() > 0 &&
                 ((IEnergySink) mTileEntity).injectEnergy(ForgeDirection.getOrientation(mSide), aVoltage, aVoltage) < aVoltage)
